feat: 로그인 관련된 타입들 따로 분리
